requirements.txt
alabaster==0.7.13
anyio==3.7.1
attrs==23.1.0
Babel==2.12.1
beautifulsoup4==4.12.2
black==23.3.0
build==0.10.0
certifi==2023.7.22
cfgv==3.3.1
charset-normalizer==2.1.1
click==8.1.6
colorama==0.4.6
coverage==5.5
cssselect==1.2.0
distlib==0.3.7
docutils==0.18.1
exceptiongroup==1.1.2
filelock==3.12.2
flake8==6.1.0
furo==2023.7.26
ghp-import==2.1.0
griffe==0.35.0
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
identify==2.5.26
idna==3.4
imagesize==1.4.1
importlib-metadata==6.8.0
iniconfig==2.0.0
Jinja2==3.1.2
lxml==4.9.3
Markdown==3.4.4
markdown-it-py==2.2.0
markdown2==2.4.10
MarkupSafe==2.1.3
mccabe==0.7.0
mdit-py-plugins==0.3.5
mdurl==0.1.2
mergedeep==1.3.4
mkdocs==1.5.2
mkdocs-autorefs==0.5.0
mkdocs-material==9.2.0
mkdocs-material-extensions==1.1.1
mkdocstrings==0.22.0
mkdocstrings-python==1.7.0
mypy-extensions==1.0.0
myst-parser==1.0.0
nodeenv==1.8.0
packaging==23.1
paginate==0.5.6
pathspec==0.11.2
platformdirs==3.10.0
pluggy==1.2.0
pre-commit==3.2.2
pycodestyle==2.11.0
pyflakes==3.1.0
Pygments==2.15.1
pymdown-extensions==10.1
pyproject_hooks==1.0.0
pyquery==2.0.0
pytest==7.2.2
pytest-asyncio==0.21.0
pytest-cov==4.0.0
python-dateutil==2.8.2
PyYAML==6.0.1
pyyaml_env_tag==0.1
readme-coverage-badger==0.1.2
readtime==3.0.0
regex==2023.6.3
requests==2.28.1
rfc3986==1.5.0
six==1.16.0
sniffio==1.3.0
snowballstemmer==2.2.0
soupsieve==2.4.1
Sphinx==6.1.3
sphinx-basic-ng==1.0.0b2
sphinx-rtd-theme==1.2.0
sphinxcontrib-applehelp==1.0.4
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.1
sphinxcontrib-jquery==4.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
toml==0.10.2
tomli==2.0.1
typing_extensions==4.7.1
urllib3==1.26.16
virtualenv==20.24.2
watchdog==3.0.0
websocket-client==1.5.1
zipp==3.16.2
